pyproject.toml
[project]
name = "a2a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "a2a-sdk[http-server]>=0.3.24",
    "uvicorn>=0.41.0",
]
